\name{avctoe00}
\alias{avctoe00}

\title{Converts a ARC/INFO binary coverage into an ESRI E00 file}
\description{
This function makes a convertion to an ESRI E00 file from a binary
coverage. 

}

\usage{avctoe00(avcdir, e00file)}

\arguments{
\item{avcdir}{The path to the binary coverage we want to convert from.}
\item{e00file}{The E00 file to be created.}
}

\value{
Returns 'NULL' on exit.
}

\seealso{e00toavc}


\keyword{file}



