\name{get.namesofcoverages}
\alias{get.namesofcoverages}

\title{Returns the names of the coverages}

\description{
Returns the name of all the files and directories under the given 
directory. We must use this function on the directory where all the coverages
are stored.
}

\usage{get.namesofcoveragtes(directory)}

\arguments{
\item{directory}{A character string with the name of the directory which 
contains the coverages.}
}

\values{
A list of character strings with the names of the files and directories
under the given one.
}

\keyword{file}
