\name{get.namesofcoverages}
\alias{get.namesofcoverages}

\title{Function for getting the names of the coverages}

\description{
This function returns the name of all the coverages under the given directory.
}

\usage{get.namesofcoverages(directory)}

\arguments{
\item{directory}{A character string with the name of the directory which 
contains the coverages.}
}

\value{
A list of character strings with the names of the files and directories
under the given one.
}

\keyword{file}
